added text to typeDef
